Extract project/group members list mutation functions into their own components so that they can be reused
